This commit was manufactured by cvs2svn to create tag 'v3-01-06'.
git-svn-id: http://root.cern.ch/svn/root/tags/v3-01-06@3892 27541ba8-7e3a-0410-8455-c3a389f83636

+++ build/package/lib/makedebfiles.sh
@@ -1,38 +1,35 @@
 #!/bin/sh -e 
 #
-# $Id: makedebfiles.sh,v 1.1 2001/04/23 14:11:47 rdm Exp $
+# $Id$
 #
 # Writes a files file entry to standard debian/root-<pkg>.files
 #
-tgtdir=$1 ; shift 
-cmndir=$1 ; shift
-prefix=$1 ; shift 
-etcdir=$1 ; shift 
-docdir=$1 ; shift 
+. build/package/lib/common.sh debian
+
+if [ $# -lt 1 ] ; then 
+    echo "$0: I need a package name - giving up"
+    exit 2
+fi
+
+# save package name in logical variable 
 pkg=$1
 
-### echo %%% List of file types to put into files 
+# List of file types to put into files
 types="shlibs conffiles files" 
 
-### echo %%% Make sure we get a fresh file 
+# Make sure we get a fresh file 
 rm -f ${tgtdir}/${pkg}.files
 
-### echo %%% for each type convert to one line
+# for each type convert to one line
 for i in $types ; do 
-    ### echo %%% Check if we have got a $i file for this $pkg in $cmndir
-    if test -f $cmndir/$pkg.$i  ; then 
-	### echo %%% Got $cmndir/$pkg.$i
+    # Check if we've got shared libraries for this package 
+    if [ -f $cmndir/$pkg.$i ] ; then 
 	grep -v -e "^#" $cmndir/$pkg.$i | \
-	    sed -e "s,@prefix@,${prefix},g" \
-		-e "s,@etcdir@,${etcdir},g" \
-		-e "s,@docdir@,${docdir},g" | \
+	    sed 's/^\/\(.*\)/\1/' | \
 	    tr "\n" " " >> ${tgtdir}/${pkg}.files
     fi 
 done 
 
 #
-# $Log: makedebfiles.sh,v $
-# Revision 1.1  2001/04/23 14:11:47  rdm
-# part of the debian and redhat build system.
-#
+# $Log$
 #
